refactor: rename all component files to follow PascalCase convention
- email-login.tsx → EmailLogin.tsx
- email-register.tsx → EmailRegister.tsx
- feedback-form.tsx → FeedbackForm.tsx
- forgot-password-*.tsx → ForgotPassword*.tsx
- password-strength-indicator.tsx → PasswordStrengthIndicator.tsx
- phone-login.tsx → PhoneLogin.tsx

- custom-dialog.tsx → CustomDialog.tsx
- image-inject.tsx → ImageInject.tsx
- profile-inject.tsx → ProfileInject.tsx
- video-inject.tsx → VideoInject.tsx

- error-display.tsx → ErrorDisplay.tsx
- terms-checkbox.tsx → TermsCheckbox.tsx
- image-uploader.tsx → ImageUploader.tsx
- solid-content.tsx → SolidContent.tsx

- profile-form-item.tsx → ProfileFormItem.tsx
- profile-save.tsx → ProfileSave.tsx
- profile-select.tsx → ProfileSelect.tsx
- profile-select-modal.tsx → ProfileSelectModal.tsx

- alert-dialog.tsx → AlertDialog.tsx
- dropdown-menu.tsx → DropdownMenu.tsx
- navigation-menu.tsx → NavigationMenu.tsx
- radio-group.tsx → RadioGroup.tsx
- scroll-area.tsx → ScrollArea.tsx
- use-alert.tsx → UseAlert.tsx

- Updated all import statements across the codebase to match new file names
- Maintained functionality while enforcing consistent PascalCase naming
