+++ src/cook-web/src/components/auth/EmailLogin.tsx
@@ -8,7 +8,8 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { useToast } from '@/hooks/use-toast';
 import { Loader2 } from 'lucide-react';
-import { TermsCheckbox } from '@/components/terms-checkbox';
+import { TermsCheckbox } from '@/components/TermsCheckbox';
+import apiService from '@/api';
 import { isValidEmail } from '@/lib/validators';
 import { useTranslation } from 'react-i18next';
 import { useAuth } from '@/hooks/useAuth';
