+++ src/cook-web/src/components/auth/PasswordStrengthIndicator.tsx
