+++ src/cook-web/src/components/auth/PhoneLogin.tsx
@@ -8,7 +8,8 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Loader2 } from 'lucide-react';
 import { useToast } from '@/hooks/use-toast';
-import { TermsCheckbox } from '@/components/terms-checkbox';
+import { TermsCheckbox } from '@/components/TermsCheckbox';
+import apiService from '@/api';
 import { isValidPhoneNumber } from '@/lib/validators';
 import { useTranslation } from 'react-i18next';
 import i18n from '@/i18n';
