+++ src/cook-web/src/components/profiles/ProfileFormItem.tsx
@@ -4,7 +4,7 @@ import type React from 'react';
 import { useState, useEffect } from 'react';
 import { Badge } from '@/components/ui/badge';
 import { X } from 'lucide-react';
-import ProfileSelectModal from './profile-select-modal';
+import ProfileSelectModal from './ProfileSelectModal';
 import useProfiles from './useProfiles';
 import type { Profile } from '@/components/profiles/type';
